pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
# PyPI distribution name (CLI commands remain `crossfire` / `crossfire-proxy`).
# The unqualified name `crossfire` is taken on PyPI by another project.
name = "crossfire-mcp"
version = "0.1.1"
description = "MCP & A2A security proxy with real-time threat detection"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Crossfire contributors" }]
keywords = ["mcp", "security", "proxy", "a2a", "ai-safety", "threat-detection", "json-rpc"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
]
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.34",
  "websockets>=14.0",
  "pyyaml>=6.0",
  "python-Levenshtein>=0.26",
  "httpx>=0.28",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
]
gemini = [
  "google-adk>=1.0",
  "google-generativeai>=0.8",
]

[project.urls]
Homepage = "https://github.com/Yugandhar-G/crossfire"
Repository = "https://github.com/Yugandhar-G/crossfire"
Documentation = "https://github.com/Yugandhar-G/crossfire#readme"

[project.scripts]
crossfire = "proxy.__main__:main"
crossfire-proxy = "proxy.proxy_stdio_main:main"

[tool.setuptools.packages.find]
include = ["proxy*", "server*", "shared*"]

[tool.setuptools.package-data]
server = ["web_dist/**/*"]

[tool.setuptools]
include-package-data = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
